Move UpdateIssuesCommit from models to repofiles
